+++ CHANGELOG.md
@@ -28,6 +28,8 @@ if you wish to upgrade.
   [`Arrays.hasMultiple()`](https://cengels.github.io/potence/Arrays/hasMultiple)
 - **Added**: new function
   [`Arrays.difference()`](https://cengels.github.io/potence/Arrays/difference)
+- **Added**: new function
+  [`Arrays.intersection()`](https://cengels.github.io/potence/Arrays/intersection)
 - **Renamed**: `Arrays.inBounds()` to
   [`Arrays.isInBounds()`](https://cengels.github.io/potence/Arrays/isInBounds)
 - **Renamed**: `Arrays.empty()` to